LayoutTests/platform/chromium/fast/forms/time/time-suggestion-picker-mouse-operations.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/common.js"></script>
<script src="../../../../../fast/forms/resources/suggestion-picker-common.js"></script>
</head>
<body style="background-color: #bbbbbb;">
<p id="description"></p>
<div id="console"></div>
<input type=time id=time value="01:00" list=suggestions>
<datalist id=suggestions>
    <option label="Now">01:01</option>
    <option>01:02</option>
    <option>01:03</option>
    <option>01:04</option>
    <option>01:05</option>
    <option>01:06</option>
    <option>01:07</option>
    <option>01:08</option>
    <option>01:09</option>
    <option>01:10</option>
    <option>01:11</option>
    <option>01:12</option>
    <option>01:13</option>
    <option>01:14</option>
    <option>01:15</option>
    <option>01:16</option>
    <option>01:17</option>
    <option>01:18</option>
    <option>01:19</option>
    <option>01:20</option>
    <option>01:21</option>
</datalist>
<script>
debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

window.onload = function() {
    openPicker(document.getElementById('time'));
    shouldBe('popupWindow.pagePopupController.toString()', '"[object PagePopupController]"');

    popupWindow.addEventListener("resize", function() {
        debug('Check that page popup exists.');
        shouldBe('popupWindow.pagePopupController.toString()', '"[object PagePopupController]"');

        debug('Check that hovering over an entry highlights it.');
        hoverOverElement(popupWindow.document.querySelector(".suggestion-list-entry:nth-child(2)"));
        shouldBe('highlightedEntry()', '"01:02"');

        debug('Check that moving the mouse outside the popup de-highlights entries.');
        hoverOverElement(document.getElementById('description'));
        shouldBeNull('highlightedEntry()');

        debug('Check that mouse click closes the popup and sets the value.');
        clickElement(popupWindow.document.querySelector(".suggestion-list-entry:nth-child(3)"));
        shouldBeNull('document.getElementById("mock-page-popup")');
        shouldBe('document.getElementById("time").value', '"01:03"');

        finishJSTest();
    }, false);
};
</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
